update item and supplier entity

+++ Backend/jeewalokadigital/jeewalokadigital/src/main/java/com/jeewaloka/digital/jeewalokadigital/repository/ItemRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemRepo extends JpaRepository<Item,Integer> {
+public interface ItemRepo extends JpaRepository<Item,Long> {
 }
